split out data_providers into a module
